Moved embedded timestamp screens here.
emb-timestamps.edl and emb-tpr-timestamps.edl have been
cloned and tweaked in various ioc/common/*Cam IOCs.
It makes more sense to have them here in timeStampFifo
as most of the PVs in these screens are specific to
this module.   Putting them both together makes it easier
for future developers to keep them uptodate with
new enhancements to timeStampFifo.